+++ fish_speech/inference_engine/reference_loader.py
@@ -8,7 +8,7 @@ import torchaudio
 from loguru import logger
 
 from fish_speech.models.vqgan.modules.firefly import FireflyArchitecture
-from tools.file import AUDIO_EXTENSIONS, audio_to_bytes, list_files, read_ref_text
+from fish_speech.utils.file import AUDIO_EXTENSIONS, audio_to_bytes, list_files, read_ref_text
 from tools.schema import ServeReferenceAudio
