+++ tools/api_client.py
@@ -8,7 +8,7 @@ import requests
 from pydub import AudioSegment
 from pydub.playback import play
 
-from tools.file import audio_to_bytes, read_ref_text
+from fish_speech.utils.file import audio_to_bytes, read_ref_text
 from tools.schema import ServeReferenceAudio, ServeTTSRequest
